refactor: renombrar posts/PostDetail por posts/DetailHeader, mover posts/components/DetailComtent a posts; renombrar posts/componenets/PostDetailProfileImage por PostDetailHeaderImage; renombrar posts/componenets/PostDetailHeader por PostDetailHeaderInfo
