+++ components/features/posts/post-detail-header.tsx
@@ -2,23 +2,21 @@ import { Button } from "@/components/ui/button";
 import { PostView } from "@/interfaces/post/post-view.interface";
 import { FC } from "react";
 
-import { PostDetailDescription } from "./components/post-detail-description";
-import { PostDetailtHeader } from "./components/post-detail-header";
-import { PostDetailMedia } from "./components/post-detail-media";
-import { PostDetailProfileImage } from "./components/post-detail-profile-image";
+import { PostDetailtHeaderInfo } from "./components/post-detail-header-info";
 import MoreIcon from "./icons/more";
+import { PostDetailHeaderImage } from "./components/post-detail-header-image";
 
-export const PostDetail: FC<PostView> = (post) => {
+export const PostDetailHeader: FC<PostView> = (post) => {
   const { urlProfile, name, date, description, media } = post;
 
   return (
     <>
       <div className="overflow-hidden break-words pt-1">
-        <PostDetailProfileImage urlProfile={urlProfile} />
+        <PostDetailHeaderImage urlProfile={urlProfile} />
       </div>
       <div className="overflow-hidden break-words">
         <div className="flex items-center justify-between">
-          <PostDetailtHeader name={name} date={date} />
+          <PostDetailtHeaderInfo name={name} date={date} />
           <Button
             variant="ghost"
             size="icon"
